fyp-backend/pyproject.toml
[project]
name = "fyp-backend"
version = "0.1.0"
description = "FastAPI backend skeleton"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "FYP Team" }]
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<0.32",
    "python-dotenv>=1.0,<2.0",
    "pydantic-settings>=2.5,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "pymysql>=1.1,<2.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "bcrypt>=4.0,<4.1",
    "python-jose[cryptography]>=3.3,<4.0",
    "email-validator>=2.1,<3.0",
    "rapidfuzz>=3.9,<4.0",
    "numpy>=1.26,<3.0",
    "scikit-learn>=1.4,<2.0",
    "joblib>=1.3,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "httpx>=0.27,<0.28",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
